+++ ErabliereIU/src/admin/admin-erablieres/erabliere-list/erabliere-list.component.ts
@@ -1,6 +1,6 @@
 import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {Erabliere} from "../../../model/erabliere";
-import {AdminErabliereAccessListComponent} from "../../../access/admin-erabliere-access-list.component";
+import {AdminErabliereAccessListComponent} from "../../../access/erabliere-access-list/admin-erabliere-access-list.component";
 import {CustomerAccess} from "../../../model/customerAccess";
 
 @Component({
