+++ ErabliereIU/src/erablieres/modifier-erabliere.component.html
@@ -26,7 +26,7 @@
                         </button>
                         } @else {
                         <h5>Accès</h5>
-                    <access-list [idErabliere]="erabliere?.id"></access-list>
+                    <erabliere-access-list [idErabliere]="erabliere?.id"></erabliere-access-list>
                     }
                 </div>
                 }
